sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intentprotocol"
version = "0.1.0"
description = "Build AI agents that negotiate and transact — Intent Protocol SDK"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Intent Protocol Contributors"},
]
keywords = ["intent", "protocol", "agent", "negotiation", "ai-agents", "marketplace"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: AsyncIO",
]

dependencies = [
    "websockets>=12.0",
    "pynacl>=1.5.0",
]

[project.urls]
Homepage = "https://intentprotocol.org"
Repository = "https://github.com/intentprotocol/sdk-python"

[tool.setuptools.packages.find]
include = ["intentprotocol*"]
